Module diploid_contam_estimator::contamination_estimator

Functions

calaulate_loglik_for_variant_position 🔒

Helper function to calculate the log probability of a given variant

calc_loglik_for_hypothetical_contam_level

Calculate log probability of seeing a number of alt calls at some read depth for a given contamination level

calc_loglik_for_hypothetical_contam_level_heterozygous 🔒

Return log probability of a heterozygous variant for a given contamination level for heterozygous variant, it is a little more complex, because it could be due to:

calculate_contam_hypothesis

Given a list of variant position and a hypothetical contamination level we calculate the log probabilty of seeing the given numbers of alt base across all positions