Function diploid_contam_estimator::contamination_estimator::calaulate_loglik_for_variant_position

fn calaulate_loglik_for_variant_position(
    variant_position: &VariantPosition,
    hypothetical_contamination_level: f64
) -> Result<Hypothesis, String>

Helper function to calculate the log probability of a given variant

Arguments

variant_position: A VariantPosition object to be evaluated

hypothetical_contamination_level: hypothetical contamination level to test

Returns

the best hypothesis with the highest log probability of seeing the given variant alt coutn