Function diploid_contam_estimator::contamination_estimator::calc_loglik_for_hypothetical_contam_level

pub fn calc_loglik_for_hypothetical_contam_level(
    variant_position: &VariantPosition, 
    hypothetical_contamination_level: f64
) -> f64

Calculate log probability of seeing a number of alt calls at some read depth for a given contamination level

Arguments

variant_position: a VariantPosition instance

hypothetical_contamination_level: a conamination level to test for

Returns

log probability of seeing the given number of alt calls

Examples

use assert_approx_eq::assert_approx_eq;
use diploid_contam_estimator::contamination_estimator::calc_loglik_for_hypothetical_contam_level;
use diploid_contam_estimator::model::{VariantPosition, VariantType, Zygosity};
let variant = VariantPosition::new(
    "chr1", 1, 50, 25, VariantType::SNV, Zygosity::HETEROZYGOUS
).unwrap();
let log_prob = calc_loglik_for_hypothetical_contam_level(&variant, 0.2);
assert_approx_eq!(-13.343980087895773, log_prob);