Function diploid_contam_estimator::contamination_estimator::calculate_contam_hypothesis

pub fn calculate_contam_hypothesis(
    variant_list: &mut Vec<VariantPosition>,
    hypothetical_contamination_level: f64
) -> Result<f64, String>

Given a list of variant position and a hypothetical contamination level we calculate the log probabilty of seeing the given numbers of alt base across all positions

Arguments

variant_list - a list of VariantPosition

hypothetical_contamination_level - the hypthetical contamination level

Returns

the sum of log probabilty of seeing the given list of variants at the given contam level

Examples

use assert_approx_eq::assert_approx_eq;
use diploid_contam_estimator::contamination_estimator::calculate_contam_hypothesis;
use diploid_contam_estimator::model::{VariantPosition, VariantType, Zygosity};
let contam_level: f64 = 0.0;
let expected_log_prob: f64 = -2.5308764039;
let mut variant_list: Vec<VariantPosition> = vec![
    VariantPosition::new("X", 1, 100, 50, VariantType::SNV, Zygosity::HETEROZYGOUS).unwrap(),
    VariantPosition::new("X", 1, 100, 100, VariantType::SNV, Zygosity::HOMOZYGOUS).unwrap(),
];
let log_prob: f64 = calculate_contam_hypothesis(&mut variant_list, contam_level).unwrap();
assert_approx_eq!(log_prob, expected_log_prob)