Function diploid_contam_estimator::run

pub fn run(
    vcf_file: &str,
    loci_bed: Option<&str>,
    snv_only_flag: bool,
    depth_threshold: usize,
    prob_json: Option<&str>,
    variant_json: Option<&str>
) -> Result<f64, String>

the actual workflow to takes in a variant vcf file and calcualte the contamination level

Arguments:

vcf_file: the file path to the input vcf file for the analysis

snv_only_flag: boolean flag indicating whether we should only look at SNV instead of both SNV and indel

depth_threshold: removing all variants with read depth below this threshold

prob_json: for debug, a json file name for writing the contam level and the respecitive log likelihoos into (“_no_file” will turn off writing a file)

prob_json: for debug, a json file name for writing the list of variants that are being used for the contam level compuatation (“_no_file” will turn off writing a file)

Return:

the contamination level with the highest log likelihood

Examples:

use diploid_contam_estimator::run;
let best_guess_contam = run("data/test.vcf", None, true, 100, Some("prob.json"), Some("variant.json")).unwrap();
assert_eq!(best_guess_contam, 0.046);