Function diploid_contam_estimator::vcfreader::build_variant_list

pub fn build_variant_list(
    vcf_file: &str,
    snv_only_flag: bool,
    depth_threshold: usize,
    regions: Vec<String>
) -> Result<Vec<VariantPosition>, String>

Colelcting variants from a vcf file

Arguments:

vcf_file: file path to the vcf file we want to parse

snv_only_flag: boolean flag indicating whether we shopuld only look at SNV instead of both SNV and indel

depth_threshold: we will skip any variants with DP tag lower than this threshold

Returns:

a list of variants that passed the given filters

Examples

use diploid_contam_estimator::vcfreader::build_variant_list;
let variant_list = build_variant_list("data/test.vcf", true, 100, vec![]).unwrap();
assert_eq!(variant_list.len(), 7);