Function diploid_contam_estimator::vcfreader::filter_variants

fn filter_variants(
    record: &Record,
    depth_threshold: usize,
    snv_only_flag: bool
) -> Option<VariantPosition>

Evaluate a vcf record and determine whether it should be collected for estimating contamination

Arguments

record: a vcf record from noodles_vcf

variants: a mutable list containing the accepted variants

depth_threshold: if the variant has DP tag lower than this, it will be rejected

snv_only_flag: boolean flag indicating whether we should skip all InDel variants

Return

VariantPosition object or a Option::None if the variant didn’t pass the filters